requirements.txt
streamlit==1.32.0
openai-whisper==20231117
yt-dlp==2024.3.10
ffmpeg-python==0.2.0

requests
